create function to view all users from database

+++ src/main/java/com/example/manifest/repository/ManifestRepository.java
@@ -3,5 +3,5 @@ package com.example.manifest.repository;
 import com.example.manifest.Entity.Manifest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ManifestRepository extends JpaRepository<Manifest,Long>{
+public interface ManifestRepository extends JpaRepository<Manifest,Integer>{
 }
